backend/legal_engine/requirements.txt

fastapi==0.109.0
uvicorn==0.27.0
Pillow==10.2.0
python-bidi==0.4.2
arabic-reshaper==3.0.0
requests==2.31.0
